speed/requirements.txt
prettytable
speedtest-cli
SQLAlchemy
psycopg2-binary
sqlalchemy-utils
pytest
pytest-cov
prometheus_client==0.3.1